# [REVIEW] Refactor `memory_usage` to improve performance