--- a/python/cudf/cudf/core/frame.py
+++ b/python/cudf/cudf/core/frame.py
@@ -344,7 +344,9 @@ class Frame(BinaryOperand, Scannable):
                 "The deep parameter is ignored and is only included "
                 "for pandas compatibility."
             )
-        return {name: col.memory_usage for name, col in self._data.items()}
+        return list(self._data.names), [
+            col.memory_usage for col in self._data.columns
+        ]
 
     def __len__(self):
         return self._num_rows