--- a/python/cudf/cudf/core/index.py
+++ b/python/cudf/cudf/core/index.py
@@ -914,7 +914,7 @@ class GenericIndex(SingleColumnFrame, BaseIndex):
 
     @_cudf_nvtx_annotate
     def memory_usage(self, deep=False):
-        return sum(super().memory_usage(deep=deep).values())
+        return sum(super().memory_usage(deep=deep)[1])
 
     @_cudf_nvtx_annotate
     def equals(self, other, **kwargs):