@vyasr 2022-03-30
Thanks for that catch @bdice.

@galipremsagar could you post new benchmarks so that we know what we're getting out of this?