@galipremsagar 2022-03-30
> @galipremsagar could you post new benchmarks so that we know what we're getting out of this?

```python
In [1]: import cudf

In [2]: df = cudf.DataFrame({'a':[1, 2, 3], 'b':['a', 'b', 'c'], 'd':[111, 123, 123]})

In [3]: df = df.set_index(['a', 'd'])

In [4]: x = df.index

# branch-22.06

In [5]: %timeit x.memory_usage()
12 µs ± 48.4 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [6]: %timeit df.memory_usage()
232 µs ± 1.55 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)



# Now
In [7]: %timeit x.memory_usage()
10.3 µs ± 70.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)


In [8]: %timeit df.memory_usage()
222 µs ± 4.89 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```