@galipremsagar 2022-03-30
Discussed offline with @vyasr and decided to optimize `sizeof_cudf_dataframe`:


```python
In [4]: df = cudf.DataFrame({'a':[1, 2, 3], 'b':['a', 'b', 'c'], 'd':[111, 123, 123]})

In [5]: %timeit df.memory_usage()
203 µs ± 2.37 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [6]: %timeit dask_cudf.backends.sizeof_cudf_dataframe(df)
1.8 µs ± 14 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```